Add tiktoken library for tokenization and update git_diff_context function
The tiktoken library is added to the project to provide tokenization functionality.
The get_token_length function is added to helpers.py to get the number of tokens in a string using the tiktoken library.
The git_diff_context function is updated to include new files in the diff output and to provide 10 lines of context.
Tests are added for the new functions.