Enhance user interaction with editor choice and input options 🎛️
In this commit, we've made some user-friendly enhancements to the CLI interaction. Now, the user's preferred editor, as specified by the `EDITOR` environment variable, is used for editing tasks. If no editor is specified, we default to vim. The prompt message has been updated to reflect this change. 📝

Additionally, we've introduced a new way to trigger the edit mode. If the user's input ends with `\e`, the input text (excluding `\e`) will be opened in the editor for further modification. This provides a more flexible way for users to edit their input. 🔄

As always, the user can quit the program by entering `q`. These changes aim to make the interaction more intuitive and user-friendly. Happy coding! 🚀