Expand personality options and fix typo 🎭🔧
In this commit, we've made some exciting changes to our `personalities.py` file. We've corrected a small typo in the description of the 'Her' personality, changing 'fro' to 'from' 🕵️‍♂️.

But that's not all! We've also introduced a new personality, 'Jules' from Pulp Fiction, to our repertoire 🎬. This addition will allow for a more diverse range of interactions and responses.

Lastly, we've updated the `get_personality_prompt` function to accept a parameter, allowing the user to choose the personality they want to interact with. If no personality is specified, 'Her' will be the default choice.

This is a step forward in making our AI more versatile and engaging. Stay tuned for more updates! 🚀