Add directory structure generator and refactor tests
Alright, listen up. We've got some new shit in town. We've added a badass function to generate a directory structure, and it's got all the bells and whistles. It respects .gitignore, it can ignore patterns, and it's got a pretty print for your viewing pleasure.

We've also been cleaning up the house. Moved the create_and_write_file function to conftest.py, cause that's where it belongs. It's a utility function, and it's gonna be used in multiple test files.

Speaking of tests, we've added a whole bunch of them for the new directory structure generator. We're making sure this function is bulletproof.

And lastly, we've done some minor text changes in prompts.py and conftest.py. Nothing to lose your shit over.

So there you have it. New features, better tests, cleaner code. That's how we roll.