Refine README to reflect current capabilities and future plans
The README has been updated to more accurately represent the current state and future trajectory of AICodeBot. The changes include a more precise description of tasks AICodeBot can handle, tasks that are in progress, and tasks that will be possible as the technology evolves. The aim is to provide a clearer picture to users about what AICodeBot can and cannot do at this point in time, and what they can expect in the future.