Use file type info and modification time when sending file context
In the 'prompts.py' file, the method of generating file context has been refactored to include more detailed information about each file, such as whether it's binary and when it was last modified. The 'arrow' library has been added to facilitate this change. This library has also been added to the 'requirements.in' and 'requirements.txt' files.