Update Dockerfile and README.md
The Dockerfile has been updated to include the configuration for the user email and name. This will ensure that the correct user information is associated with the commits made by the Docker container.

The README.md file has been updated to include instructions for running the Docker container assuming there are changes in the current working directory that have already been added. Additionally, an ugly hack has been added to run LLM locally.

These changes improve the functionality and usability of the project.