We recommend starting the commit message with a blank line, followed by an introduction to the changes being made. The summary should be less than 72 characters.
The body of the commit message should include the changes being made,
and any specific instructions that may be needed to get started.

If there is only one specific change, it can be a short summary. If there are multiple
changes, each one should be summarized and listed separately.

In addition to the sample commit message, we recommend using
Git tags to describe the specific changes made in each release.

Git tags are labels applied to Git objects that mark a specific version of a release.

In the case of the example commit message, you can replace the prefix with an
example of a Git tag, such as v1.0 or v1.2.