Refine token size calculation in sidekick function
Fixes #67

The previous implementation of token size calculation in the sidekick function was simplified. This commit introduces a more accurate
calculation by rounding the memory token size and creating a new function, calc_response_token_size, to dynamically calculate the response
token size based on the files. This change also includes a small buffer for the response token size and a recalculation of the response
token size in case the files change. This should lead to more precise token size management and potentially improve the performance of the
sidekick function. 🚀📊