Update Dockerfile and README.md
The Dockerfile has been updated to include additional dependencies and configurations for the aicodebot application. The README.md file has also been modified to include instructions for building and running the Docker image.

This commit improves the Dockerfile by adding necessary packages and setting up the environment for the aicodebot application. It also updates the README.md file to provide clear instructions on how to build and run the Docker image.

The changes in the Dockerfile and README.md files will make it easier for users to set up and run the aicodebot application in a Docker environment.

Note: The commit message has been generated based on the provided diff context.